refactor!: move synchronous array operations into traits
