+++ zarrs/src/array/codec/array_to_bytes/sharding.rs
@@ -58,6 +58,8 @@ mod sharding_partial_decoder_async;
 mod sharding_partial_decoder_sync;
 mod sharding_partial_encoder;
 
+pub(crate) use sharding_partial_decoder_sync::ShardingPartialDecoder;
+
 use std::borrow::Cow;
 use std::num::NonZeroU64;
 use std::sync::Arc;
@@ -72,7 +74,6 @@ pub use sharding_codec_builder::ShardingCodecBuilder;
 pub use sharding_options::{ShardingCodecOptions, SubchunkWriteOrder};
 #[cfg(feature = "async")]
 pub(crate) use sharding_partial_decoder_async::AsyncShardingPartialDecoder;
-pub(crate) use sharding_partial_decoder_sync::ShardingPartialDecoder;
 use zarrs_codec::{
     ArrayCodecTraits, ArrayToBytesCodecTraits, BytesPartialDecoderTraits, Codec, CodecError,
     CodecOptions, CodecPluginV3, CodecTraitsV3,
